+++ server/infra/lib/service-stack.ts
@@ -57,6 +57,7 @@ export interface ServiceStackProps extends StackProps {
 export class ServiceStack extends Stack {
   public readonly fargateService: FargateService
   public readonly migrationLambda: NodejsFunction
+  public readonly albFargate: ApplicationLoadBalancedFargateService
   constructor(scope: Construct, id: string, props: ServiceStackProps) {
     super(scope, id, props)
 
@@ -69,10 +70,12 @@ export class ServiceStack extends Stack {
       props.dbSecretArn,
     )
 
+    const groqApiKeyName = `${stageName}/ito/groq-api-key`
+
     const groqApiKeySecret = Secret.fromSecretNameV2(
       this,
       'GroqApiKey',
-      'groq-api-key',
+      groqApiKeyName,
     )
 
     const fargateTaskRole = new IamRole(this, 'ItoFargateTaskRole', {
@@ -96,8 +99,8 @@ export class ServiceStack extends Stack {
       'ItoTaskDefinition',
       {
         taskRole: fargateTaskRole,
-        cpu: 256,
-        memoryLimitMiB: 512,
+        cpu: 1024,
+        memoryLimitMiB: 2048,
         runtimePlatform: {
           operatingSystemFamily: OperatingSystemFamily.LINUX,
           cpuArchitecture: CpuArchitecture.ARM64,
@@ -180,6 +183,15 @@ export class ServiceStack extends Stack {
       unhealthyThresholdCount: 5,
     })
 
+    const scalableTarget = fargateService.service.autoScaleTaskCount({
+      minCapacity: 1,
+      maxCapacity: 5,
+    })
+
+    scalableTarget.scaleOnCpuUtilization('ItoServerCpuScalingPolicy', {
+      targetUtilizationPercent: 65,
+    })
+
     // Setup migration lambda
     const migrationLambda = new NodejsFunction(this, 'ItoMigrationLambda', {
       functionName: `${stageName}-${DB_NAME}-migration`,
@@ -227,6 +239,7 @@ export class ServiceStack extends Stack {
     alb.logAccessLogs(logBucket, 'ito-alb-access-logs')
 
     this.fargateService = fargateService.service
+    this.albFargate = fargateService
     this.migrationLambda = migrationLambda
 
     new CfnOutput(this, 'ServiceURL', {