+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the electron-react-app (ERA) are listed here.
 
 <br>
 
+### v0.2.3
+- Update referral sources
+- Enhance copy button tooltip to stay visible and show "Copied 🎉" feedback
+- Create about page
+- Update layout to new design
+- Self host auth0 callback page [(#58)](https://github.com/heyito/ito/pull/58)
+- Better handle errors from transcription service [(#62)](https://github.com/heyito/ito/pull/62)
+
 ### v0.2.2
 
 - check localhost, not grpc_base_url